fix: combine styleguidist options plugin tests into single file
