+++ src/scripts/utils/__tests__/StyleguidistOptionsPlugin.spec.ts
@@ -6,15 +6,16 @@ const options: any = {
 };
 const mockContext: { _styleguidist?: Rsg.StyleguidistConfig } = {};
 
+let mockedModule: Record<string, unknown>;
+
 jest.mock('webpack', () => {
-	const resource = { resource: 'test' };
 	return {
 		NormalModule: {
 			getCompilationHooks: () => {
 				return {
 					loader: {
 						tap: (moduleName: string, compilationCallback: (context: any, opt: any) => void) => {
-							compilationCallback(mockContext, resource);
+							compilationCallback(mockContext, mockedModule);
 						},
 					},
 				};
@@ -23,7 +24,24 @@ jest.mock('webpack', () => {
 	};
 });
 
+it('should do nothing when module.resource is not present', () => {
+	mockedModule = {};
+	const compiler = {
+		hooks: {
+			compilation: {
+				tap: (name: string, callback: (opt: any) => void) => {
+					callback({});
+				},
+			},
+		},
+	};
+	const plugin = new StyleguidistOptionsPlugin(options);
+	plugin.apply(compiler as any);
+	expect(mockContext._styleguidist).toBeFalsy();
+});
+
 it('should attach Styleguidist config options', () => {
+	mockedModule = { resource: 'test' };
 	const compiler = {
 		hooks: {
 			compilation: {
